Conflicts:
	CHANGELOG.md
	src/Pim/Bundle/CatalogBundle/Version.php

+++ src/Pim/Bundle/CatalogBundle/Manager/GroupManager.php
@@ -2,11 +2,14 @@
 
 namespace Pim\Bundle\CatalogBundle\Manager;
 
+use Pim\Bundle\CatalogBundle\Repository\ProductRepositoryInterface;
+use Pim\Bundle\CatalogBundle\Model\GroupInterface;
+use Pim\Bundle\CatalogBundle\Event\GroupEvents;
+use Akeneo\Component\StorageUtils\Remover\RemoverInterface;
+use Pim\Bundle\CatalogBundle\Repository\AttributeRepositoryInterface;
+use Symfony\Bridge\Doctrine\RegistryInterface;
 use Symfony\Component\EventDispatcher\EventDispatcherInterface;
 use Symfony\Component\EventDispatcher\GenericEvent;
-use Symfony\Bridge\Doctrine\RegistryInterface;
-use Pim\Bundle\CatalogBundle\Event\GroupEvents;
-use Pim\Bundle\CatalogBundle\Entity\Group;
 
 /**
  * Group manager
@@ -15,7 +18,7 @@ use Pim\Bundle\CatalogBundle\Entity\Group;
  * @copyright 2013 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
-class GroupManager
+class GroupManager implements RemoverInterface
 {
     /** @var RegistryInterface */
     protected $doctrine;
@@ -35,36 +38,42 @@ class GroupManager
     /** @var string */
     protected $attributeClass;
 
+    /** @var ProductRepositoryInterface */
+    protected $productRepository;
+
     /**
      * Constructor
      *
-     * @param RegistryInterface        $doctrine
-     * @param EventDispatcherInterface $eventDispatcher
-     * @param string                   $groupClass
-     * @param string                   $groupTypeClass
-     * @param string                   $productClass
-     * @param string                   $attributeClass
+     * @param RegistryInterface          $doctrine
+     * @param EventDispatcherInterface   $eventDispatcher
+     * @param ProductRepositoryInterface $productRepository
+     * @param string                     $groupClass
+     * @param string                     $groupTypeClass
+     * @param string                     $productClass
+     * @param string                     $attributeClass
      */
     public function __construct(
         RegistryInterface $doctrine,
         EventDispatcherInterface $eventDispatcher,
+        ProductRepositoryInterface $productRepository,
         $groupClass,
         $groupTypeClass,
         $productClass,
         $attributeClass
     ) {
-        $this->doctrine        = $doctrine;
-        $this->eventDispatcher = $eventDispatcher;
-        $this->groupClass      = $groupClass;
-        $this->groupTypeClass  = $groupTypeClass;
-        $this->productClass    = $productClass;
-        $this->attributeClass  = $attributeClass;
+        $this->doctrine          = $doctrine;
+        $this->eventDispatcher   = $eventDispatcher;
+        $this->groupClass        = $groupClass;
+        $this->groupTypeClass    = $groupTypeClass;
+        $this->productClass      = $productClass;
+        $this->attributeClass    = $attributeClass;
+        $this->productRepository = $productRepository;
     }
 
     /**
      * Get available axis
      *
-     * @return \Pim\Bundle\CatalogBundle\Model\AbstractAttribute[]
+     * @return \Pim\Bundle\CatalogBundle\Model\AttributeInterface[]
      */
     public function getAvailableAxis()
     {
@@ -80,7 +89,7 @@ class GroupManager
     {
         $attributes = $this->getAvailableAxis();
 
-        $choices = array();
+        $choices = [];
         foreach ($attributes as $attribute) {
             $choices[$attribute->getId()] = $attribute->getLabel();
         }
@@ -111,9 +120,9 @@ class GroupManager
      */
     public function getTypeChoices($isVariant)
     {
-        $types = $this->getGroupTypeRepository()->findBy(array('variant' => $isVariant));
+        $types = $this->getGroupTypeRepository()->findBy(['variant' => $isVariant]);
 
-        $choices = array();
+        $choices = [];
         foreach ($types as $type) {
             $choices[$type->getId()] = $type->getLabel();
         }
@@ -143,62 +152,46 @@ class GroupManager
     }
 
     /**
-     * Removes a group
-     *
-     * @param Group $group
+     * {@inheritdoc}
      */
-    public function remove(Group $group)
+    public function remove($group, array $options = [])
     {
+        if (!$group instanceof GroupInterface) {
+            throw new \InvalidArgumentException(
+                sprintf('Expects a "Pim\Bundle\CatalogBundle\Model\GroupInterface", "%s" provided.', get_class($group))
+            );
+        }
+
         $this->eventDispatcher->dispatch(GroupEvents::PRE_REMOVE, new GenericEvent($group));
 
+        $options = array_merge(['flush' => true], $options);
         $em = $this->doctrine->getManager();
         $em->remove($group);
-        $em->flush();
+        if (true === $options['flush']) {
+            $em->flush();
+        }
     }
 
     /**
      * Returns an array containing a limited number of product groups, and the total number of products
      *
-     * @param Group   $group
-     * @param integer $maxResults
+     * @param GroupInterface $group
+     * @param integer        $maxResults
      *
      * @return array
      */
-    public function getProductList(Group $group, $maxResults)
+    public function getProductList(GroupInterface $group, $maxResults)
     {
-        $manager = $this->doctrine->getManager();
-        $products = $manager
-            ->createQueryBuilder()
-            ->select('p')
-            ->from($this->productClass, 'p')
-            ->innerJoin('p.groups', 'g', 'WITH', 'g=:group')
-            ->setParameter('group', $group)
-            ->getQuery()
-            ->setMaxResults($maxResults + 1)
-            ->execute();
+        $products = $this->productRepository->getProductsByGroup($group, $maxResults);
+        $count = $this->productRepository->getProductCountByGroup($group);
 
-        $count = count($products);
-        if ($count > $maxResults) {
-            array_pop($products);
-            $count = $manager->createQueryBuilder()
-                ->select('COUNT(p)')
-                ->from($this->productClass, 'p')
-                ->innerJoin('p.groups', 'g', 'WITH', 'g=:group')
-                ->setParameter('group', $group)
-                ->getQuery()
-                ->getSingleScalarResult();
-        }
-
-        return array(
-            'products'      => $products,
-            'productCount'  => $count
-        );
+        return ['products' => $products, 'productCount' => $count];
     }
 
     /**
      * Get the attribute repository
-     *
-     * @return \Pim\Bundle\CatalogBundle\Entity\Repository\AttributeRepository
+
+     * @return AttributeRepositoryInterface
      */
     protected function getAttributeRepository()
     {
